Create Calculate the Factorial of a Number with For cycle
Write a program that can find the factorial of any given number. For example, find
the factorial of the number 5 (often written as 5!) which is 1 * 2 * 3 * 4 * 5 and
equals 120.
The factorial is not defined for negative numbers and the factorial of Zero is 1;
that is 0! = 1.
84 7 Iteration/Looping
Your program should take as input an integer from the user (you can reuse your
logic from the last chapter to verify that they have entered a positive integer value
using isnumeric()).
You should
1. If the number is less than Zero return with an error message.
2. Check to see if the number is Zero—if it is then the answer is 1—print this out.
3. Otherwise use a loop to generate the result and print it out.